Fixed alignment issues

+++ MQ2DanNet/deps/archive/archive.h
@@ -337,7 +337,7 @@ class Archive
 
         const Archive& operator&(const std::string& v) const
         {
-            size_t len = v.length();
+            uint32_t len = static_cast<uint32_t>(v.length());
             *this & len;
             m_stream.write(v.c_str(), len);
             return *this;
